backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.119.0",
    "faster-whisper>=1.2.0",
    "ffmpeg-python>=0.2.0",
    "openai>=2.4.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.37.0",
]
